requirements.txt
click==6.6
Flask==0.11.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.11.10
wheel==0.24.0
